wip: omnibar
