+++ packages/frontend/src/features/omnibar/index.ts
@@ -1,0 +1,2 @@
+import Omnibar from './components/Omnibar';
+export default Omnibar;
